requirements-dev.txt
# Development dependencies
black==23.7.0
isort==5.12.0
flake8==6.1.0
mypy==1.5.1
pylint==2.17.5
pydocstyle==6.3.0
bandit==1.7.5
safety==2.3.5
pre-commit==3.3.3
pytest-cov==4.1.0
pytest-xdist==3.3.1
pytest-rerunfailures==11.1.2
pytest-metadata==3.0.0
pytest-html==4.1.1
allure-pytest==2.13.2
pytest-asyncio==0.21.1
pytest-timeout==2.1.0
pytest-env==1.1.1
pytest-mock==3.11.1
pytest-instafail==0.4.2
pytest-sugar==0.9.7
pytest-bdd==6.1.1
pytest-html-reporter==0.2.9
pytest-selenium==4.1.0
pytest-selenium-snapshot==0.4.0
pytest-snapshot==0.9.0
pytest-split==0.8.1
pytest-testrail==3.2.0
pytest-tldr==0.2.4
pytest-tornado==0.8.1
pytest-vagrant==0.2.0
pytest-watch==4.2.0
pytest-xvfb==2.0.0
pytest-yapf3==1.2.0
pytest-yaml==0.1.0
pytest-json-report==1.5.0
pytest-json==0.4.0
pytest-json-schema==1.4.0
pytest-json-report-wip==0.1.0
pytest-json-snapshot==0.1.0
pytest-jsonrpc==0.1.0
pytest-jsonrpc2==0.1.0
pytest-jsonrpc-client==0.1.0
pytest-jsonrpc-server==0.1.0
pytest-jsonrpc2-client==0.1.0
pytest-jsonrpc2-server==0.1.0
pytest-jsonrpc2-transport-http==0.1.0
pytest-jsonrpc2-transport-ws==0.1.0
pytest-jsonrpc2-transport==0.1.0
pytest-jsonrpc2==0.1.0
pytest-jsonrpc-client==0.1.0
pytest-jsonrpc-server==0.1.0
pytest-jsonrpc2-client==0.1.0
pytest-jsonrpc2-server==0.1.0
pytest-jsonrpc2-transport-http==0.1.0
pytest-jsonrpc2-transport-ws==0.1.0
pytest-jsonrpc2-transport==0.1.0
pytest-jsonrpc2==0.1.0
